+++ javascript-modules/engines/astro-engine/lib/modules/assets.js
@@ -1,0 +1,5 @@
+import ImageInternal from './image.astro';
+import PictureInternal from './picture.astro';
+
+export const Image = ImageInternal;
+export const Picture = PictureInternal;